Code-review R2: parser test-coverage holes (T-2..T-6)
Second PR of the 2026-07-11 remediation series — closes the cheap
parser/coverage gaps from the review doc. Test-only; no production
logic changes.

- T-2: the duplicate-clause-keyword branch in clause_bounds.rs had zero
  coverage. Unit tests pin the message and second-occurrence caret for
  DIMENSIONS and TABLES; end-to-end sqllogictest asserts the CREATE
  pipeline rejects it.
- T-3: pin empty-clause-body behaviour (TABLES()/DIMENSIONS()/METRICS()/
  MATERIALIZATIONS()) at both the parse-unit layer and end-to-end — the
  'at least one of D/M' rule is about clause PRESENCE, so both-empty is a
  legal degenerate parse; downstream expansion, not the scanner, owns the
  zero-table rejection.
- T-4: comments inside the AS body work only by construction (whole-query
  blanking runs before the body parser). Unit + sqllogictest cover block
  comments between clauses and inside parens; a unit test pins that a typo
  after an in-body comment still carets at its true byte offset
  (length-preserving blanking).
- T-6: wildcard interaction cells — alias.* expansion when the view holds
  a semi-additive metric (active snapshot path) and a window metric (the
  window+aggregate mix must surface through wildcard expansion), plus
  mix-free dims-wildcard + explicit window metric.
- T-5: factor the hostile-identifier/payload proptest generators into
  tests/common/mod.rs and add create_front_door_proptest.rs, which drives
  the FULL plan_rewrite CREATE route (comment-blanking, prefix detection,
  name extraction, offset threading) with rendered hostile defs — the
  fixed-body CREATE proptest never varied the body. The shared generator
  now also matches FK arity to the target PK so every def is valid for the
  full pipeline (inference populates omitted ref_columns from the PK), a
  strict improvement that keeps the round-trip property intact.

Verification: cargo test (1094 lib + all integration incl. the new
256-case front-door proptest), sqllogictest 69/69, DuckLake CI, fmt,
clippy (default + --features extension, both -D warnings).
